requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Flask==2.1.1
Flask-Login==0.6.0
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
prompt-toolkit==3.0.29
protobuf==3.20.0
pycurl==7.45.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.3
requests==2.27.1
rq==1.10.1
sib-api-v3-sdk==7.4.0
six==1.16.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.1
zipp==3.8.0
